added 10e and 5e evaluation, extended evaluation to question sources

+++ llama/src/03_train_lora.py
@@ -22,7 +22,7 @@ bnb_config = BitsAndBytesConfig(
     load_in_4bit=True,
     bnb_4bit_use_double_quant=True,
     bnb_4bit_quant_type="nf4",
-    bnb_4bit_compute_dtype=torch.float16
+    bnb_4bit_compute_dtype=torch.bfloat16
 )
 
 # log into huggingface hub
